+++ mindspore/ops/kernel/gpu/cuda/nn/adagrad_v2_gpu_kernel.cc
@@ -37,7 +37,8 @@ void AdagradV2GpuKernelMod::InOutputResize(const std::vector<KernelTensor *> &in
                                                                   inputs[kIndex1]->GetDeviceShapeVector().end());
   std::vector<int64_t> gradient_shape_ = std::vector<int64_t>(inputs[kIndex3]->GetDeviceShapeVector().begin(),
                                                               inputs[kIndex3]->GetDeviceShapeVector().end());
-  input_elements_ = std::accumulate(variable_shape_.begin(), variable_shape_.end(), 1, std::multiplies<int64_t>());
+  input_elements_ = std::accumulate(variable_shape_.begin(), variable_shape_.end(), static_cast<int64_t>(1),
+                                    std::multiplies<int64_t>());
 
   is_null_input_ = (input_elements_ == 0);
